Add Scala.js bindings

+++ core/src/main/scala/org/polyvariant/treesitter4s/TreeSitter.scala
@@ -34,6 +34,6 @@ trait Node {
   def tpe: String
   def children: List[Node]
   def fields: Map[String, Node]
-  def startByte: Long
-  def endByte: Long
+  def startByte: Int
+  def endByte: Int
 }
